refactor(constants.ts): Move all `typings` into `constants.ts`
