--- typings/BidSubmitType.ts
@@ -1,7 +1,0 @@
-export interface IBidSubmitData {
-  date: string;
-  time: string;
-  volume: string;
-  price: string;
-  total_price: string;
-}
