--- typings/ConnectedAMI.ts
@@ -1,7 +1,0 @@
-import { PowerType } from './PowerType';
-
-export interface IConnectedAMI {
-  num: number;
-  id: string;
-  type: PowerType;
-}
