--- typings/PowerType.ts
@@ -1,7 +1,0 @@
-export enum PowerType {
-  PV = '太陽能',
-  Homepage = '淨負載',
-  WT = '風能',
-  ESS = '儲能系統',
-  EV = '充電樁',
-}
